feat: add deep-research report workflow + fact-check workflow + builtin skill
Two complementary research workflows:

- deep-research (NEW): comprehensive report generator.
  Brief → Plan → parallel sub-agent research → Reflect gap-check →
  single-writer Markdown report → cold review + fix. Supports
  quick/standard/deep depth budgets. Convergent: file checkpoints
  (brief.md / plan.json / F*.md / reflect.json / REPORT.md) enable
  resume after interruption.

- fact-check (RENAMED from deep-research): adversarial fact verifier.
  Search → extract checkable facts → group duplicates → 3-juror
  adversarial crosscheck → structured JSON findings. Best for precise
  claims ("Is X true?").

Also adds deep-research builtin skill (.bundle) with sub-agent prompt
template, free source endpoint reference, and report format spec.

Sandbox quirks addressed: args arrives as JSON string (JSON.parse);
Date unavailable (caller passes args.today); no hardcoded paths.
